fix: post폴더 수정 및 백기능 테스트 완료
