+++ blog-service/build.gradle
@@ -1,3 +1,4 @@
+
 plugins {
     id 'java'
     id 'org.springframework.boot' version '3.5.3'
@@ -25,6 +26,7 @@ repositories {
 
 dependencies {
     implementation 'net.coobird:thumbnailator:0.4.14'
+    annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
 
     implementation 'org.modelmapper:modelmapper:3.2.1'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
@@ -44,3 +46,4 @@ dependencies {
 tasks.named('test') {
     useJUnitPlatform()
 }
+
