+++ blog-service/src/main/java/com/playblog/blogservice/post/entity/Post.java
@@ -1,4 +1,4 @@
-package com.playblog.blogservice.postservice.post.entity;
+package com.playblog.blogservice.post.entity;
 
 import com.playblog.blogservice.common.entity.SubTopic;
 import com.playblog.blogservice.common.entity.TopicType;
@@ -15,9 +15,12 @@ import com.playblog.blogservice.user.User;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Post {
     /* 정책 참조 */
-    private Boolean allowComment; // 댓글 허용
-    private Boolean allowLike;    // 좋아요 허용
-    private Boolean allowSearch;  // 검색 허용
+    @Column(nullable = false)
+    private Boolean allowComment;
+    @Column(nullable = false)
+    private Boolean allowLike;
+    @Column(nullable = false)
+    private Boolean allowSearch;
 
     /* 유저 참조 */
     @ManyToOne(fetch = FetchType.LAZY)
