+++ blog-service/src/main/java/com/playblog/blogservice/search/repository/SearchRepository.java
@@ -1,7 +1,7 @@
 package com.playblog.blogservice.search.repository;
 
 import com.playblog.blogservice.common.entity.SubTopic;
-import com.playblog.blogservice.postservice.post.entity.Post;
+import com.playblog.blogservice.post.entity.Post;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
