+++ blog-service/src/main/java/com/playblog/blogservice/search/service/SearchService.java
@@ -5,7 +5,7 @@ import com.playblog.blogservice.common.entity.TopicType;
 import com.playblog.blogservice.common.exception.ErrorCode;
 import com.playblog.blogservice.common.exception.SearchException;
 import com.playblog.blogservice.neighbor.Repository.NeighborRepository;
-import com.playblog.blogservice.postservice.post.entity.Post;
+import com.playblog.blogservice.post.entity.Post;
 
 import com.playblog.blogservice.search.dto.*;
 import com.playblog.blogservice.search.repository.*;
